First version of the map generator
First version of the map generator. The main problems are fractal-like generation of rooms, which undermines the impression of randomly generated map. The probable reason is the way it checks how a new room can be attached - it always starts from one direction, which is usually avaible. Also, the increasing number of doors not attached to the spawn makes a map look like a tree. I will add something which will make "more than one door out the spawn room" scenario more prevalent. Lastly, after the generation, every door which is not used should be deleted, so a player cannot go out-of-bounds.